pyproject.toml
[tool.poetry]
package-mode = false
name = "e-commerce"
version = "0.1.0"
description = "ECommerce platform built with Python and Django."
authors = ["Yousef Abu Shanab <josephyousef249@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/youzarsiph/e-commerce"
repository = "https://github.com/youzarsiph/e-commerce"
packages = [{ include = "e_commerce" }]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.10.3"
Django = "^5.1"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
drf-redesign = "^0.4.0"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
huggingface-hub = "^0.25.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.5.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
